feat: Add support for specifying the Port in the connection string
Before this it was impossible to run multiple AMQP-based service emulators on the same host with different port mappings, e.g. by running each emulator in a Docker container with the internal 5672 port mapped to an arbitrary host port. Now with the ability to specify the port we can create such a set up.